+++ CHANGELOG.md
@@ -3,10 +3,16 @@
 All notable changes to this project will be documented in this file.
 
 
+## Version 1.9.73
+
+- Added ref_shard_blocks to ShardStateUnsplit
+- Added new parameter 'collators' to ShardHashes::add_workchain
+- Add milliseconds to state
+- Added end_lt to ShardBlockRef
+
 ## Version 1.9.68
 
 - Added BlockInfo::gen_utime_ms (#1)
-- Prepare to open tests
 - Open library tests
 - Fix compiler warnings
 - Increase package version
@@ -18,7 +24,6 @@ All notable changes to this project will be documented in this file.
 ## Version 1.9.63
 
 - Added "collator" and "ref shard blocks" fields for Venom
-
 
 ## Version 1.9.47